autobyteus-image-audio/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autobyteus-image-audio"
version = "0.1.0"
description = "MCP server exposing Autobyteus image and audio generation tools"
authors = [{name = "Autobyteus"}]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "mcp>=1.13.1",
    "autobyteus==1.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-anyio>=0",
    "trio>=0.23",
]

[project.scripts]
autobyteus-image-audio-server = "image_audio_mcp.server:main"

[tool.pytest.ini_options]
addopts = "-q"
